// src/shared/types.ts
type PromptVarValue = string | number | boolean | string[]

interface PromptVar {
  id: string
  name: string
  type: "text" | "number" | "select" | "multi-select" | "toggle"
  label?: string
  options?: string[]
  default?: PromptVarValue
}

interface SvgAvatar {
  type: "svg"
  name: string
  hue?: number
  title?: string
}
interface TextAvatar {
  type: "text"
  text: string
  hue?: number
  title?: string
}
interface ImageAvatar {
  type: "image"
  imageId: string
  hue?: number
  title?: string
}
interface UrlAvatar {
  type: "url"
  url: string
  hue?: number
  title?: string
}
interface IconAvatar {
  type: "icon"
  icon: string
  hue?: number
  title?: string
}
type Avatar = SvgAvatar | TextAvatar | ImageAvatar | UrlAvatar | IconAvatar

const ProviderSchema = Object({
  type: String(),
  settings: Object(undefined),
})
type Provider = Static<typeof ProviderSchema>
interface ProviderType {
  name: string
  label: string
  avatar: Avatar
  settings: PluginSchema
  initialSettings
  constructor: (settings) => any
  getModelList?: (settings) => Promise<string[]>
}

interface ApiResultItem {
  type: "text" | "file" | "quote"
  contentText?: string
  contentBuffer?: ArrayBuffer
  name?: string
  mimeType?: string
}
// ...
interface FileparserData {
  enabled: boolean
  mimeTypes: string[]
}

interface AssistantPluginAction {
  name: string
  enabled: boolean
  autoExecute: boolean
}

interface AssistantPlugin {
  enabled: boolean
  infos: AssistantPluginInfo[]
  tools: AssistantPluginTool[]
  resources: AssistantPluginResource[]
  vars: Record<string, PromptVarValue>
}

interface AssistantPlugins {
  [id: string]: AssistantPlugin
}

interface PluginApi {
  type: "info" | "tool"
  infoType?: "resource" | "prompt" | "prompt-var"
  name: string
  description?: string
  prompt?: string
  parameters: PluginSchema
  showComponents?: string[]
  execute(args, settings): Promise<ApiResultItem[]>
}

interface PluginFileparser {
  name: string
  description: string
  rangeInput?: {
    label?: string
    hint?: string
    mask?: string
  }
  execute(
    args: { file: Blob; range?: string },
    settings
  ): Promise<ApiResultItem[]>
}

interface Plugin {
  id: string
  type: "builtin" | "lobechat" | "gradio" | "mcp"
  available: boolean
  apis: PluginApi[]
  fileparsers: PluginFileparser[]
  title: string
  description?: string
  prompt?: string
  promptVars?: PromptVar[]
  settings: PluginSchema
  noRoundtrip?: boolean
  author?: string
  homepage?: string
}
// ...
interface PluginData {
  settings
  avatar: Avatar
  fileparsers: Record<string, FileparserData>
}